[29] Content of test suite 5 moved to test suite 4.
