+++ tests/test_suite4_parent_paths.py
@@ -6,12 +6,14 @@ import sys
 from _test_utils import\
 	INIT_SYS_PATH,\
 	REPO_ROOT,\
+	assert_path_is_present,\
 	index_in_sys_path,\
 	reset_sys_path
 
 sys.path.insert(0, str(REPO_ROOT))
 from syspathmodif import\
-	sp_prepend_parent
+	sp_prepend_parent,\
+	sp_prepend_parent_bundle
 reset_sys_path()
 
 
@@ -58,5 +60,37 @@ def test_prepend_parent_index_error():
 	try:
 		with pytest.raises(IndexError):
 			sp_prepend_parent(2025)
+	finally:
+		reset_sys_path()
+
+
+def test_parent_bundle():
+	try:
+		bundle = sp_prepend_parent_bundle((0, 1, 2, -1))
+
+		parent0 = _THIS_FILE.parents[0]
+		assert_path_is_present(parent0, bundle, True, False)
+		assert index_in_sys_path(parent0) == 3
+
+		parent1 = _THIS_FILE.parents[1]
+		assert_path_is_present(parent1, bundle, True, True)
+		assert index_in_sys_path(parent1) == 2
+
+		parent2 = _THIS_FILE.parents[2]
+		assert_path_is_present(parent2, bundle, True, True)
+		assert index_in_sys_path(parent2) == 1
+
+		parent_minus1 = _THIS_FILE.parents[-1]
+		assert_path_is_present(parent_minus1, bundle, True, True)
+		assert index_in_sys_path(parent_minus1) == 0
+
+	finally:
+		reset_sys_path()
+
+
+def test_prepend_bundle_index_error():
+	try:
+		with pytest.raises(IndexError):
+			sp_prepend_parent_bundle((0, 1, 2, 2025))
 	finally:
 		reset_sys_path()
